pkg_tar_runfiles: demonstration of issue packaging an application + runfiles
1. Build the pkg_tar rule intended to package the application:

	bazel build path/to/some/app:app_deploy

2. See the resulting file:

	tar -tf bazel-bin/path/to/some/app/app_deploy.tar

It is desirable that included data files are at their workspace-relative
path. For example,

	./path/to/some/app/public/stylesheets/app.css
	./shared/libraries/some_lib/file.txt

However, the observed behavior is that filegroups defined in the same BUILD file
are at a relative path to that package:

	./public/stylesheets/app.css

while other files are at their workspace-relative path:

	./shared/libraries/some_lib/file.txt

This seems like a surprising combination which makes it hard to load files in
development and production.